+++ lib/fastererer/formatters.rb
@@ -3,6 +3,7 @@
 require_relative 'formatters/text_formatter'
 require_relative 'formatters/json_formatter'
 require_relative 'formatters/rdjsonl_formatter'
+require_relative 'formatters/github_formatter'
 
 module Fastererer
   class UnknownFormatError < StandardError; end
@@ -11,7 +12,8 @@ module Fastererer
     FORMATS = {
       'text' => TextFormatter,
       'json' => JsonFormatter,
-      'rdjsonl' => RdjsonlFormatter
+      'rdjsonl' => RdjsonlFormatter,
+      'github' => GithubFormatter
     }.freeze
 
     def self.fetch(name)